+++ posti.config.example.js
@@ -8,14 +8,19 @@ module.exports.default = {
   dialectOptions: {}, // http://docs.sequelizejs.com/manual/installation/usage.html#options
 
   app: {
-    chunkSize: 1000, // In how big chunks do we want to insert the data into database.
-    concurrency: 5, // Number of concurrent inserts.
-    deleteOnComplete: true, // Should the temporary data directory be removed after script finishes.
+    // In how big chunks do we want to insert the data into database.
+    chunkSize: 1000,
+    // Number of concurrent inserts.
+    concurrency: 5,
+    // Should the temporary data directory be removed after script finishes.
+    deleteOnComplete: true,
   },
 
-  table: {
+  tables: {
     addresses: {
+      // Name for the database table.
       name: 'addresses',
+      // Should the script use `temp` table for updating or not.
       useTempTable: true,
     },
     postalcodes: {
@@ -25,6 +30,29 @@ module.exports.default = {
     postalcode_changes: {
       name: 'postalcode_changes',
       useTempTable: false,
+    },
+  },
+
+  server: {
+    // In which port should the GraphQL server be run in.
+    port: 3000,
+    // To enable query tracing, change this to true.
+    tracing: false,
+    // To enable caching, change this to true.
+    cacheControl: false,
+    // To disable GraphiQL and Playground, change this to true.
+    production: false,
+
+    // https://www.npmjs.com/package/express-rate-limit#configuration
+    limiter: {
+      // 5 minutes
+      windowMs: 5 * 60 * 1000,
+      // limit each IP to 250 requests per windowMs
+      max: 250,
+      // delay after n requests.
+      delayAfter: 5,
+      // disable delaying - full speed until the max limit is reached
+      delayMs: 100,
     },
   },
 };